Initial conversion to gem format
