+++ lib/jekyll/webmention_io/version.rb
@@ -1,0 +1,5 @@
+module Jekyll
+  module WebmentionIo
+    VERSION = "0.1.0"
+  end
+end
