+++ lib/jekyll/webmention_io/webmention_count_tag.rb
@@ -1,0 +1,21 @@
+#  (c) Aaron Gustafson
+#  https://github.com/aarongustafson/jekyll-webmention_io 
+#  Licence : MIT
+
+module Jekyll
+  module WebmentionIo
+    class WebmentionCountTag < Webmentions
+      
+      def initialize(tagName, text, tokens)
+        super
+        @api_endpoint = 'http://webmention.io/api/count'
+      end
+
+      def html_output_for(response)
+        count = response['count'] || '0'
+        "<span class=\"webmention-count\">#{count}</span>"
+      end
+      
+    end
+  end
+end
